Separates player importer to separate file
